backend-engineer/microservices/gateway/requirements.txt
fastapi
uvicorn
httpx
python-jose
pyjwt
